+++ src/main/kotlin/carRacing/model/Record.kt
@@ -1,4 +1,4 @@
-package carRacing
+package carRacing.model
 
 data class Record(
     val positions: List<Int>
