+++ src/main/kotlin/carRacing/util/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package carRacing
+package carRacing.util
 
 import kotlin.random.Random
