+++ src/main/kotlin/carRacing/view/InputView.kt
@@ -1,4 +1,4 @@
-package carRacing
+package carRacing.view
 
 object InputView {
